Fixed DeleteBatch
The one thing that I haven't fixed yet, is when you confirm the deletion the page should reload, so that you can see that the batch is deleted, but this doesn't happen. You still get the notification telling you that the deletion is succesfull, but the reload isn't done yet.